resize-worker/requirements.txt
pika==0.13.0
Pillow==5.4.1
